Changes logging of MS and PFS to human readable logs
